move ExpressionActionsChain logic out of InterpreterSelectQuery
