+++ dbms/src/Interpreters/InterpreterSelectQuery.cpp
@@ -154,9 +154,7 @@ String InterpreterSelectQuery::generateFilterActions(ExpressionActionsPtr & acti
     /// Using separate expression analyzer to prevent any possible alias injection
     auto syntax_result = SyntaxAnalyzer(*context).analyze(query_ast, storage->getColumns().getAllPhysical());
     SelectQueryExpressionAnalyzer analyzer(query_ast, syntax_result, *context);
-    ExpressionActionsChain new_chain(*context);
-    analyzer.appendSelect(new_chain, false);
-    actions = new_chain.getLastActions();
+    actions = analyzer.simpleSelectActions(*context);
 
     return expr_list->children.at(0)->getColumnName();
 }
@@ -210,22 +208,6 @@ static Context getSubqueryContext(const Context & context)
     subquery_settings.extremes = 0;
     subquery_context.setSettings(subquery_settings);
     return subquery_context;
-}
-
-static bool sanitizeBlock(Block & block)
-{
-    for (auto & col : block)
-    {
-        if (!col.column)
-        {
-            if (isNotCreatable(col.type->getTypeId()))
-                return false;
-            col.column = col.type->createColumn();
-        }
-        else if (isColumnConst(*col.column) && !col.column->empty())
-            col.column = col.column->cloneEmpty();
-    }
-    return true;
 }
 
 InterpreterSelectQuery::InterpreterSelectQuery(
@@ -556,11 +538,18 @@ Block InterpreterSelectQuery::getSampleBlockImpl(bool try_move_to_prewhere)
     if (storage && !options.only_analyze)
         from_stage = storage->getQueryProcessingStage(*context);
 
-    analysis_result = analyzeExpressions(
+    /// Do I need to perform the first part of the pipeline - running on remote servers during distributed processing.
+    bool first_stage = from_stage < QueryProcessingStage::WithMergeableState
+        && options.to_stage >= QueryProcessingStage::WithMergeableState;
+    /// Do I need to execute the second part of the pipeline - running on the initiating server during distributed processing.
+    bool second_stage = from_stage <= QueryProcessingStage::WithMergeableState
+        && options.to_stage > QueryProcessingStage::WithMergeableState;
+
+    analysis_result = ExpressionAnalysisResult(
             getSelectQuery(),
             *query_analyzer,
-            from_stage,
-            options.to_stage,
+            first_stage,
+            second_stage,
             *context,
             storage,
             options.only_analyze,
@@ -614,261 +603,6 @@ Block InterpreterSelectQuery::getSampleBlockImpl(bool try_move_to_prewhere)
     }
 
     return analysis_result.final_projection->getSampleBlock();
-}
-
-/// Check if there is an ignore function. It's used for disabling constant folding in query
-///  predicates because some performance tests use ignore function as a non-optimize guard.
-static bool allowEarlyConstantFolding(const ExpressionActions & actions, const Context & context)
-{
-    if (!context.getSettingsRef().enable_early_constant_folding)
-        return false;
-
-    for (auto & action : actions.getActions())
-    {
-        if (action.type == action.APPLY_FUNCTION && action.function_base)
-        {
-            auto name = action.function_base->getName();
-            if (name == "ignore")
-                return false;
-        }
-    }
-    return true;
-}
-
-
-void InterpreterSelectQuery::AnalysisResult::finalize(const ExpressionActionsChain & chain, const Context & context_, size_t where_step_num)
-{
-    if (hasPrewhere())
-    {
-        const ExpressionActionsChain::Step & step = chain.steps.at(0);
-        prewhere_info->remove_prewhere_column = step.can_remove_required_output.at(0);
-
-        Names columns_to_remove;
-        for (size_t i = 1; i < step.required_output.size(); ++i)
-        {
-            if (step.can_remove_required_output[i])
-                columns_to_remove.push_back(step.required_output[i]);
-        }
-
-        if (!columns_to_remove.empty())
-        {
-            auto columns = prewhere_info->prewhere_actions->getSampleBlock().getNamesAndTypesList();
-            ExpressionActionsPtr actions = std::make_shared<ExpressionActions>(columns, context_);
-            for (const auto & column : columns_to_remove)
-                actions->add(ExpressionAction::removeColumn(column));
-
-            prewhere_info->remove_columns_actions = std::move(actions);
-        }
-
-        columns_to_remove_after_prewhere = std::move(columns_to_remove);
-    }
-    else if (hasFilter())
-    {
-        /// Can't have prewhere and filter set simultaneously
-        filter_info->do_remove_column = chain.steps.at(0).can_remove_required_output.at(0);
-    }
-    if (hasWhere())
-        remove_where_filter = chain.steps.at(where_step_num).can_remove_required_output.at(0);
-}
-
-void InterpreterSelectQuery::AnalysisResult::removeExtraColumns()
-{
-    if (hasFilter())
-        filter_info->actions->prependProjectInput();
-    if (hasWhere())
-        before_where->prependProjectInput();
-    if (hasHaving())
-        before_having->prependProjectInput();
-}
-
-void InterpreterSelectQuery::AnalysisResult::checkActions()
-{
-    /// Check that PREWHERE doesn't contain unusual actions. Unusual actions are that can change number of rows.
-    if (hasPrewhere())
-    {
-        auto check_actions = [](const ExpressionActionsPtr & actions)
-        {
-            if (actions)
-                for (const auto & action : actions->getActions())
-                    if (action.type == ExpressionAction::Type::JOIN || action.type == ExpressionAction::Type::ARRAY_JOIN)
-                        throw Exception("PREWHERE cannot contain ARRAY JOIN or JOIN action", ErrorCodes::ILLEGAL_PREWHERE);
-        };
-
-        check_actions(prewhere_info->prewhere_actions);
-        check_actions(prewhere_info->alias_actions);
-        check_actions(prewhere_info->remove_columns_actions);
-    }
-}
-
-
-InterpreterSelectQuery::AnalysisResult
-InterpreterSelectQuery::analyzeExpressions(
-    const ASTSelectQuery & query,
-    SelectQueryExpressionAnalyzer & query_analyzer,
-    QueryProcessingStage::Enum from_stage,
-    QueryProcessingStage::Enum to_stage,
-    const Context & context,
-    const StoragePtr & storage,
-    bool only_types,
-    const FilterInfoPtr & filter_info,
-    const Block & source_header)
-{
-    AnalysisResult res;
-
-    /// Do I need to perform the first part of the pipeline - running on remote servers during distributed processing.
-    res.first_stage = from_stage < QueryProcessingStage::WithMergeableState
-        && to_stage >= QueryProcessingStage::WithMergeableState;
-    /// Do I need to execute the second part of the pipeline - running on the initiating server during distributed processing.
-    res.second_stage = from_stage <= QueryProcessingStage::WithMergeableState
-        && to_stage > QueryProcessingStage::WithMergeableState;
-
-    /** First we compose a chain of actions and remember the necessary steps from it.
-        *  Regardless of from_stage and to_stage, we will compose a complete sequence of actions to perform optimization and
-        *  throw out unnecessary columns based on the entire query. In unnecessary parts of the query, we will not execute subqueries.
-        */
-
-    bool finalized = false;
-    size_t where_step_num = 0;
-
-    auto finalizeChain = [&](ExpressionActionsChain & chain)
-    {
-        if (!finalized)
-        {
-            chain.finalize();
-            res.finalize(chain, context, where_step_num);
-            chain.clear();
-        }
-        finalized = true;
-    };
-
-    {
-        ExpressionActionsChain chain(context);
-        Names additional_required_columns_after_prewhere;
-
-        if (storage && (query.sample_size() || context.getSettingsRef().parallel_replicas_count > 1))
-        {
-            Names columns_for_sampling = storage->getColumnsRequiredForSampling();
-            additional_required_columns_after_prewhere.insert(additional_required_columns_after_prewhere.end(),
-                columns_for_sampling.begin(), columns_for_sampling.end());
-        }
-
-        if (storage && query.final())
-        {
-            Names columns_for_final = storage->getColumnsRequiredForFinal();
-            additional_required_columns_after_prewhere.insert(additional_required_columns_after_prewhere.end(),
-                columns_for_final.begin(), columns_for_final.end());
-        }
-
-        if (storage && filter_info)
-        {
-            res.filter_info = filter_info;
-            query_analyzer.appendPreliminaryFilter(chain, filter_info->actions, filter_info->column_name);
-        }
-
-        if (query_analyzer.appendPrewhere(chain, !res.first_stage, additional_required_columns_after_prewhere))
-        {
-            res.prewhere_info = std::make_shared<PrewhereInfo>(
-                    chain.steps.front().actions, query.prewhere()->getColumnName());
-
-            if (allowEarlyConstantFolding(*res.prewhere_info->prewhere_actions, context))
-            {
-                Block before_prewhere_sample = source_header;
-                if (sanitizeBlock(before_prewhere_sample))
-                {
-                    res.prewhere_info->prewhere_actions->execute(before_prewhere_sample);
-                    auto & column_elem = before_prewhere_sample.getByName(query.prewhere()->getColumnName());
-                    /// If the filter column is a constant, record it.
-                    if (column_elem.column)
-                        res.prewhere_constant_filter_description = ConstantFilterDescription(*column_elem.column);
-                }
-            }
-            chain.addStep();
-        }
-
-        res.need_aggregate = query_analyzer.hasAggregation();
-
-        query_analyzer.appendArrayJoin(chain, only_types || !res.first_stage);
-
-        if (query_analyzer.appendJoin(chain, only_types || !res.first_stage))
-        {
-            res.before_join = chain.getLastActions();
-            if (!res.hasJoin())
-                throw Exception("No expected JOIN", ErrorCodes::LOGICAL_ERROR);
-            chain.addStep();
-        }
-
-        if (query_analyzer.appendWhere(chain, only_types || !res.first_stage))
-        {
-            where_step_num = chain.steps.size() - 1;
-            res.before_where = chain.getLastActions();
-            if (allowEarlyConstantFolding(*res.before_where, context))
-            {
-                Block before_where_sample;
-                if (chain.steps.size() > 1)
-                    before_where_sample = chain.steps[chain.steps.size() - 2].actions->getSampleBlock();
-                else
-                    before_where_sample = source_header;
-                if (sanitizeBlock(before_where_sample))
-                {
-                    res.before_where->execute(before_where_sample);
-                    auto & column_elem = before_where_sample.getByName(query.where()->getColumnName());
-                    /// If the filter column is a constant, record it.
-                    if (column_elem.column)
-                        res.where_constant_filter_description = ConstantFilterDescription(*column_elem.column);
-                }
-            }
-            chain.addStep();
-        }
-
-        if (res.need_aggregate)
-        {
-            query_analyzer.appendGroupBy(chain, only_types || !res.first_stage);
-            query_analyzer.appendAggregateFunctionsArguments(chain, only_types || !res.first_stage);
-            res.before_aggregation = chain.getLastActions();
-
-            finalizeChain(chain);
-
-            if (query_analyzer.appendHaving(chain, only_types || !res.second_stage))
-            {
-                res.before_having = chain.getLastActions();
-                chain.addStep();
-            }
-        }
-
-        bool has_stream_with_non_joned_rows = (res.before_join && res.before_join->getTableJoinAlgo()->hasStreamWithNonJoinedRows());
-        res.optimize_read_in_order =
-            context.getSettingsRef().optimize_read_in_order
-            && storage && query.orderBy()
-            && !query_analyzer.hasAggregation()
-            && !query.final()
-            && !has_stream_with_non_joned_rows;
-
-        /// If there is aggregation, we execute expressions in SELECT and ORDER BY on the initiating server, otherwise on the source servers.
-        query_analyzer.appendSelect(chain, only_types || (res.need_aggregate ? !res.second_stage : !res.first_stage));
-        res.selected_columns = chain.getLastStep().required_output;
-        res.has_order_by = query_analyzer.appendOrderBy(chain, only_types || (res.need_aggregate ? !res.second_stage : !res.first_stage), res.optimize_read_in_order);
-        res.before_order_and_select = chain.getLastActions();
-        chain.addStep();
-
-        if (query_analyzer.appendLimitBy(chain, only_types || !res.second_stage))
-        {
-            res.before_limit_by = chain.getLastActions();
-            chain.addStep();
-        }
-
-        query_analyzer.appendProjectResult(chain);
-        res.final_projection = chain.getLastActions();
-
-        finalizeChain(chain);
-    }
-
-    /// Before executing WHERE and HAVING, remove the extra columns from the block (mostly the aggregation keys).
-    res.removeExtraColumns();
-
-    res.subqueries_for_sets = query_analyzer.getSubqueriesForSets();
-
-    res.checkActions();
-    return res;
 }
 
 static Field getWithFillFieldValue(const ASTPtr & node, const Context & context)
