+++ dbms/src/Interpreters/InterpreterSelectQuery.h
@@ -152,62 +152,6 @@ private:
     template <typename TPipeline>
     void executeImpl(TPipeline & pipeline, const BlockInputStreamPtr & prepared_input, std::optional<Pipe> prepared_pipe, QueryPipeline & save_context_and_storage);
 
-    struct AnalysisResult
-    {
-        bool need_aggregate = false;
-        bool has_order_by   = false;
-
-        bool remove_where_filter = false;
-        bool optimize_read_in_order = false;
-
-        ExpressionActionsPtr before_join;   /// including JOIN
-        ExpressionActionsPtr before_where;
-        ExpressionActionsPtr before_aggregation;
-        ExpressionActionsPtr before_having;
-        ExpressionActionsPtr before_order_and_select;
-        ExpressionActionsPtr before_limit_by;
-        ExpressionActionsPtr final_projection;
-
-        /// Columns from the SELECT list, before renaming them to aliases.
-        Names selected_columns;
-
-        /// Columns will be removed after prewhere actions execution.
-        Names columns_to_remove_after_prewhere;
-
-        /// Do I need to perform the first part of the pipeline - running on remote servers during distributed processing.
-        bool first_stage = false;
-        /// Do I need to execute the second part of the pipeline - running on the initiating server during distributed processing.
-        bool second_stage = false;
-
-        SubqueriesForSets subqueries_for_sets;
-        PrewhereInfoPtr prewhere_info;
-        FilterInfoPtr filter_info;
-        ConstantFilterDescription prewhere_constant_filter_description;
-        ConstantFilterDescription where_constant_filter_description;
-
-        bool hasFilter() const { return filter_info.get(); }
-        bool hasJoin() const { return before_join.get(); }
-        bool hasPrewhere() const { return prewhere_info.get(); }
-        bool hasWhere() const { return before_where.get(); }
-        bool hasHaving() const { return before_having.get(); }
-        bool hasLimitBy() const { return before_limit_by.get(); }
-
-        void removeExtraColumns();
-        void checkActions();
-        void finalize(const ExpressionActionsChain & chain, const Context & context, size_t where_step_num);
-    };
-
-    static AnalysisResult analyzeExpressions(
-        const ASTSelectQuery & query,
-        SelectQueryExpressionAnalyzer & query_analyzer,
-        QueryProcessingStage::Enum from_stage,
-        QueryProcessingStage::Enum to_stage,
-        const Context & context,
-        const StoragePtr & storage,
-        bool only_types,
-        const FilterInfoPtr & filter_info,
-        const Block & source_header);
-
     /** From which table to read. With JOIN, the "left" table is returned.
      */
     static void getDatabaseAndTableNames(const ASTSelectQuery & query, String & database_name, String & table_name, const Context & context);
@@ -291,7 +235,7 @@ private:
     SelectQueryInfo query_info;
 
     /// Is calculated in getSampleBlock. Is used later in readImpl.
-    AnalysisResult analysis_result;
+    ExpressionAnalysisResult analysis_result;
     FilterInfoPtr filter_info;
 
     QueryProcessingStage::Enum from_stage = QueryProcessingStage::FetchColumns;
